refactor: split typePlan, printEstimates and the share validator along their seams
Three pure-logic modules sat over the 500-code-line limit. typePlan hands its
font measurement to typeMeasurer; printEstimates hands the label-tab material
terms to printLabelTabVolume and the pattern reductions to printPatternSavings;
api/lib/validation hands the drawer guards to drawerValidation, the linked
design validator to sharedDesignsValidation, and the shared limits and error
shape to shareConstraints. Public names re-export from their original modules.

whats-new/entries.ts joins the max-lines exemption list as a pure data file.
