+++ api/lib/validation.ts
@@ -6,106 +6,22 @@
 import { sanitizeString } from './sanitize.js';
 export { sanitizeString } from './sanitize.js';
 import { isNumber, isObject, inRange, validationError } from './validationUtils.js';
-import { validateDesignerShare } from './designerValidation.js';
-import { validateAssemblyEnvelope, validateAssemblyStructure } from './assemblyValidation.js';
-
-// Constraints for shared layouts
-const SHARE_CONSTRAINTS = {
-  MAX_SIZE_BYTES: 500 * 1024, // 500KB
-  MAX_BINS: 2500,
-  GRID_MIN: 1,
-  GRID_MAX: 50,
-  LAYERS_MIN: 1,
-  LAYERS_MAX: 10,
-  CATEGORIES_MAX: 20,
-  NAME_MAX_LENGTH: 64,
-  LABEL_MAX_LENGTH: 24,
-  NOTES_MAX_LENGTH: 256,
-  // Mirrors client `CONSTRAINTS.MIN_BIN_HEIGHT` / `MIN_LAYER_HEIGHT` (2).
-  // Server <-> client divergence would let a peer persist height=1 that the
-  // recipient's CQRS schema rejects on the next mutation.
-  HEIGHT_MIN: 2,
-  // Mirrors client `GRID_PITCH_MM_MAX / 2` (drawerOutline.ts) — half the
-  // largest settable grid pitch, i.e. the largest legal grid shift within a
-  // custom perimeter. NOT `GRID_UNIT_MM_MAX / 2`: the pitch ceiling is 200mm.
-  GRID_SHIFT_MM_MAX: 100,
-  VALID_EXPIRATIONS: [30, 60, 90, 365] as const,
-  // Custom properties constraints
-  CUSTOM_PROPERTY_MAX_COUNT: 50,
-  CUSTOM_PROPERTY_KEY_MAX_LENGTH: 32,
-  CUSTOM_PROPERTY_VALUE_MAX_LENGTH: 256,
-  CUSTOM_PROPERTY_MAX_TOTAL_SIZE: 20480, // 20KB total per bin
-};
-
-/** Design ids are generated as `design_{timestamp}_{suffix}`; 64 leaves room to spare. */
-const DESIGN_ID_MAX_LENGTH = 64;
-
-/**
- * Budget for the `linkedDesigns` payload that rides alongside a shared layout.
- * Separate from the layout's own cap so a design-heavy share can't crowd out
- * the layout itself. A plain design serializes to ~3KB and a cutout-heavy one
- * to ~7KB, so this holds well over a hundred; imported-mesh designs (which
- * carry base64 geometry in `params.meshAssets`) are the only realistic way to
- * reach it, and those are skipped individually rather than failing the share.
- */
-const MAX_LINKED_DESIGNS = 250;
-const MAX_LINKED_DESIGNS_BYTES = 512 * 1024;
-// Per-assembly cap, matching the sync endpoint's design payload budget.
-const MAX_ASSEMBLY_DESIGN_BYTES = 100 * 1024;
-const DESIGN_NAME_MAX_LENGTH = 64;
-
-/** Reserved keys that cannot be used as custom property names */
-const RESERVED_PROPERTY_KEYS = [
-  'id',
-  'layerId',
-  'x',
-  'y',
-  'width',
-  'depth',
-  'height',
-  'clearanceHeight',
-  'category',
-  'label',
-  'notes',
-  'customProperties',
-  // Prevent prototype pollution
-  '__proto__',
-  'constructor',
-  'prototype',
-];
-
-export type ValidExpiration = (typeof SHARE_CONSTRAINTS.VALID_EXPIRATIONS)[number];
-
-interface OutlineVertexShape {
-  x: number;
-  y: number;
-  bulge?: number;
-}
-
-interface CornerCutShape {
-  kind: string;
-  size?: number;
-  r?: number;
-  w?: number;
-  d?: number;
-}
-
-interface DrawerOutlineShape {
-  vertices: OutlineVertexShape[];
-  authoring?: { kind: string; corners?: Record<string, CornerCutShape> };
-}
-
-interface DrawerShape {
-  width: number;
-  depth: number;
-  height: number;
-  fractionalEdgeX?: 'start' | 'end';
-  fractionalEdgeY?: 'start' | 'end';
-  outline?: DrawerOutlineShape;
-  gridShiftX?: number;
-  gridShiftY?: number;
-  measuredMm?: { width: number; depth: number; height?: number };
-}
+import { isValidDrawer, sanitizeDrawer } from './drawerValidation.js';
+import type { DrawerShape } from './drawerValidation.js';
+import { DESIGN_ID_MAX_LENGTH, RESERVED_PROPERTY_KEYS } from './sharedDesignsValidation.js';
+import { SHARE_CONSTRAINTS } from './shareConstraints.js';
+import type { ValidExpiration, ValidationError } from './shareConstraints.js';
+export { SHARE_CONSTRAINTS } from './shareConstraints.js';
+export type { ValidExpiration, ValidationError } from './shareConstraints.js';
+export {
+  DESIGN_ID_MAX_LENGTH,
+  RESERVED_PROPERTY_KEYS,
+  validateSharedDesigns,
+  isSharedDesignsError,
+} from './sharedDesignsValidation.js';
+export type { SharedDesignShape, SharedDesignsResult } from './sharedDesignsValidation.js';
+export { isValidDrawer, sanitizeDrawer } from './drawerValidation.js';
+export type { DrawerShape } from './drawerValidation.js';
 
 interface LayerShape {
   id: string;
@@ -168,11 +84,6 @@ export function withoutLibraryPlacement<T extends { folderId?: string }>(
 ): Omit<T, 'folderId'> {
   const { folderId: _folderId, ...rest } = layout;
   return rest;
-}
-
-export interface ValidationError {
-  code: 'VALIDATION_ERROR' | 'SIZE_LIMIT' | 'BIN_LIMIT' | 'INVALID_EXPIRATION';
-  message: string;
 }
 
 export type ValidationResult =
@@ -420,139 +331,6 @@ export function validateShareLayout(data: unknown, jsonSize: number): Validation
   };
 }
 
-/** A design travelling with a shared layout, keyed by the id its bins carry.
- *  Bin designs carry `params`; Workshop assemblies carry envelope + structure. */
-export interface SharedDesignShape {
-  id: string;
-  name: string;
-  params?: Record<string, unknown>;
-  kind?: 'assembly';
-  envelope?: Record<string, unknown>;
-  structure?: Record<string, unknown>;
-}
-
-export type SharedDesignsResult =
-  { valid: true; designs: SharedDesignShape[] } | { valid: false; error: ValidationError };
-
-/**
- * Validate the `linkedDesigns` payload accompanying a shared layout.
- *
- * Each design's params go through the same deep validator the single-design
- * designer share uses, so a layout share can't become a side door around the
- * stricter checks (mesh-asset cross-referencing in particular).
- *
- * Absent/empty is valid — layouts with no linked designs and older clients
- * both land here.
- */
-export function validateSharedDesigns(data: unknown): SharedDesignsResult {
-  if (data === undefined || data === null) return { valid: true, designs: [] };
-  if (!Array.isArray(data)) {
-    return validationError('VALIDATION_ERROR', 'linkedDesigns must be an array');
-  }
-  if (data.length > MAX_LINKED_DESIGNS) {
-    return validationError(
-      'VALIDATION_ERROR',
-      `Too many linked designs (max ${MAX_LINKED_DESIGNS})`
-    );
-  }
-
-  const designs: SharedDesignShape[] = [];
-  const seen = new Set<string>();
-  let totalBytes = 0;
-
-  for (const entry of data) {
-    if (!isObject(entry)) {
-      return validationError('VALIDATION_ERROR', 'Invalid linked design entry');
-    }
-    const { id, name, params } = entry;
-    if (typeof id !== 'string' || id.length === 0 || id.length > DESIGN_ID_MAX_LENGTH) {
-      return validationError('VALIDATION_ERROR', 'Invalid linked design id');
-    }
-    if (typeof name !== 'string') {
-      return validationError('VALIDATION_ERROR', 'Invalid linked design name');
-    }
-    // Bins share designs freely, so the client can emit the same id twice.
-    if (seen.has(id)) continue;
-
-    if (entry.kind === 'assembly') {
-      // Assemblies ride the same deep validators the sync endpoint uses, so a
-      // layout share can't become a side door around them.
-      const envelopeResult = validateAssemblyEnvelope(entry.envelope);
-      if (!envelopeResult.valid) {
-        return validationError(
-          'VALIDATION_ERROR',
-          `Invalid linked design envelope: ${envelopeResult.error.message}`
-        );
-      }
-      const structureResult = validateAssemblyStructure(entry.structure);
-      if (!structureResult.valid) {
-        return validationError(
-          'VALIDATION_ERROR',
-          `Invalid linked design structure: ${structureResult.error.message}`
-        );
-      }
-      const assemblyBytes = JSON.stringify({
-        envelope: entry.envelope,
-        structure: entry.structure,
-      }).length;
-      if (assemblyBytes > MAX_ASSEMBLY_DESIGN_BYTES) {
-        return validationError(
-          'SIZE_LIMIT',
-          `Linked assembly exceeds maximum size of ${MAX_ASSEMBLY_DESIGN_BYTES / 1024}KB`
-        );
-      }
-      totalBytes += assemblyBytes;
-      if (totalBytes > MAX_LINKED_DESIGNS_BYTES) {
-        return validationError(
-          'SIZE_LIMIT',
-          `Linked designs exceed maximum size of ${MAX_LINKED_DESIGNS_BYTES / 1024}KB`
-        );
-      }
-      seen.add(id);
-      designs.push({
-        id: sanitizeString(id, DESIGN_ID_MAX_LENGTH),
-        name: sanitizeString(name, DESIGN_NAME_MAX_LENGTH),
-        kind: 'assembly',
-        envelope: entry.envelope as Record<string, unknown>,
-        structure: entry.structure as Record<string, unknown>,
-      });
-      continue;
-    }
-
-    const paramsBytes = JSON.stringify(params ?? null).length;
-    const result = validateDesignerShare({ type: 'designer', version: 1, params }, paramsBytes);
-    if (!result.valid) {
-      return validationError('VALIDATION_ERROR', `Invalid linked design: ${result.error.message}`);
-    }
-
-    totalBytes += paramsBytes;
-    if (totalBytes > MAX_LINKED_DESIGNS_BYTES) {
-      return validationError(
-        'SIZE_LIMIT',
-        `Linked designs exceed maximum size of ${MAX_LINKED_DESIGNS_BYTES / 1024}KB`
-      );
-    }
-
-    seen.add(id);
-    designs.push({
-      id: sanitizeString(id, DESIGN_ID_MAX_LENGTH),
-      name: sanitizeString(name, DESIGN_NAME_MAX_LENGTH),
-      params: result.payload.params,
-    });
-  }
-
-  return { valid: true, designs };
-}
-
-/**
- * Narrow a {@link SharedDesignsResult} to its failure case.
- */
-export function isSharedDesignsError(
-  result: SharedDesignsResult
-): result is { valid: false; error: ValidationError } {
-  return !result.valid;
-}
-
 /**
  * Validate expiration days parameter.
  */
@@ -561,220 +339,6 @@ export function validateExpiration(days: unknown): days is ValidExpiration {
     typeof days === 'number' &&
     SHARE_CONSTRAINTS.VALID_EXPIRATIONS.includes(days as ValidExpiration)
   );
-}
-
-/** Mirrors CONSTRAINTS.MEASURED_MM_MIN/MAX in src/core/constants.ts. */
-const MEASURED_MM_MIN = 1;
-const MEASURED_MM_MAX = 5000;
-
-function isValidMeasuredMm(value: unknown): boolean {
-  if (!isObject(value)) return false;
-  const boundedMm = (v: unknown): boolean =>
-    isNumber(v) && inRange(v, MEASURED_MM_MIN, MEASURED_MM_MAX);
-  return (
-    boundedMm(value.width) &&
-    boundedMm(value.depth) &&
-    (value.height === undefined || boundedMm(value.height))
-  );
-}
-
-// Type guards
-const isValidGridShift = (value: unknown): value is number =>
-  isNumber(value) &&
-  inRange(value, -SHARE_CONSTRAINTS.GRID_SHIFT_MM_MAX, SHARE_CONSTRAINTS.GRID_SHIFT_MM_MAX);
-
-function isValidDrawer(value: unknown): value is DrawerShape {
-  if (!isObject(value)) return false;
-  if (value.outline !== undefined && !isValidDrawerOutline(value.outline)) return false;
-  if (value.measuredMm !== undefined && !isValidMeasuredMm(value.measuredMm)) return false;
-  if (value.gridShiftX !== undefined && !isValidGridShift(value.gridShiftX)) return false;
-  if (value.gridShiftY !== undefined && !isValidGridShift(value.gridShiftY)) return false;
-  return (
-    isNumber(value.width) &&
-    isNumber(value.depth) &&
-    isNumber(value.height) &&
-    value.width > 0 &&
-    value.depth > 0 &&
-    // HeightUnits (7mm each), capped at GRID_MAX so a corrupted client
-    // can't sync absurd values to other devices.
-    inRange(value.height, SHARE_CONSTRAINTS.HEIGHT_MIN, SHARE_CONSTRAINTS.GRID_MAX)
-  );
-}
-
-/** Mirrors OUTLINE_MAX_VERTICES in src/shared/utils/drawerOutline.ts. */
-const OUTLINE_MAX_VERTICES = 256;
-/** 50 units × up to 52mm grid, with slack — mirrors the client Zod schema. */
-const OUTLINE_COORD_MIN = -1;
-const OUTLINE_COORD_MAX = 2600;
-const OUTLINE_AUTHORING_KINDS = ['cells', 'corners', 'trace', 'pen'];
-
-/**
- * Structural gate for a drawer outline: vertex/coordinate/bulge
- * bounds and no self-intersection of the straight-chord approximation. The
- * geometric invariants that need the drawer's mm extent (CCW, min area,
- * in-bounds) are enforced client-side and re-checked at read time by
- * normalizeDrawerOutline — the server's job is to bound the data so a
- * hand-crafted payload can't sync junk or pathological geometry.
- */
-function isValidDrawerOutline(value: unknown): value is DrawerOutlineShape {
-  if (!isObject(value) || !Array.isArray(value.vertices)) return false;
-  const vertices = value.vertices as unknown[];
-  if (vertices.length < 3 || vertices.length > OUTLINE_MAX_VERTICES) return false;
-  for (const v of vertices) {
-    if (!isObject(v)) return false;
-    if (
-      !isNumber(v.x) ||
-      !isNumber(v.y) ||
-      !inRange(v.x, OUTLINE_COORD_MIN, OUTLINE_COORD_MAX) ||
-      !inRange(v.y, OUTLINE_COORD_MIN, OUTLINE_COORD_MAX)
-    ) {
-      return false;
-    }
-    if (v.bulge !== undefined && (!isNumber(v.bulge) || !inRange(v.bulge, -1, 1))) return false;
-  }
-  // `authoring` is non-geometric editor metadata: reject only structural
-  // garbage. An unknown kind (a NEWER client's editor) must not invalidate
-  // the layout — sanitizeDrawer drops unrecognized kinds instead.
-  if (value.authoring !== undefined) {
-    if (!isObject(value.authoring) || typeof value.authoring.kind !== 'string') return false;
-  }
-  const pts = vertices as OutlineVertexShape[];
-  return !outlineChordsSelfIntersect(pts);
-}
-
-/** O(n²) chord test — arcs approximated by their chords; good enough to bound
- * pathological payloads (exact arc checks run client-side). Detects proper
- * crossings AND degenerate touches: collinear overlaps and endpoint-on-segment
- * contacts between non-adjacent chords (mirrors segmentsTouch in
- * src/shared/utils/drawerOutline.ts). */
-function outlineChordsSelfIntersect(pts: OutlineVertexShape[]): boolean {
-  const n = pts.length;
-  const orient = (a: OutlineVertexShape, b: OutlineVertexShape, c: OutlineVertexShape): number => {
-    const v = (b.x - a.x) * (c.y - a.y) - (b.y - a.y) * (c.x - a.x);
-    return Math.abs(v) < 1e-9 ? 0 : Math.sign(v);
-  };
-  const onSegment = (
-    a: OutlineVertexShape,
-    b: OutlineVertexShape,
-    p: OutlineVertexShape
-  ): boolean =>
-    Math.min(a.x, b.x) - 1e-6 <= p.x &&
-    p.x <= Math.max(a.x, b.x) + 1e-6 &&
-    Math.min(a.y, b.y) - 1e-6 <= p.y &&
-    p.y <= Math.max(a.y, b.y) + 1e-6;
-  for (let i = 0; i < n; i++) {
-    const a = pts[i];
-    const b = pts[(i + 1) % n];
-    for (let j = i + 1; j < n; j++) {
-      if (j === i || (j + 1) % n === i || (i + 1) % n === j) continue;
-      const c = pts[j];
-      const d = pts[(j + 1) % n];
-      const o1 = orient(a, b, c);
-      const o2 = orient(a, b, d);
-      const o3 = orient(c, d, a);
-      const o4 = orient(c, d, b);
-      if (o1 !== o2 && o3 !== o4) return true;
-      if (o1 === 0 && onSegment(a, b, c)) return true;
-      if (o2 === 0 && onSegment(a, b, d)) return true;
-      if (o3 === 0 && onSegment(c, d, a)) return true;
-      if (o4 === 0 && onSegment(c, d, b)) return true;
-    }
-  }
-  return false;
-}
-
-/**
- * Rebuild the drawer field-by-field. The drawer previously passed through
- * raw, which let arbitrary junk keys survive sanitization; an explicit
- * rebuild keeps exactly the validated shape.
- */
-function sanitizeDrawer(drawer: DrawerShape): DrawerShape {
-  const out: DrawerShape = {
-    width: drawer.width,
-    depth: drawer.depth,
-    height: drawer.height,
-  };
-  if (drawer.fractionalEdgeX === 'start' || drawer.fractionalEdgeX === 'end') {
-    out.fractionalEdgeX = drawer.fractionalEdgeX;
-  }
-  if (drawer.fractionalEdgeY === 'start' || drawer.fractionalEdgeY === 'end') {
-    out.fractionalEdgeY = drawer.fractionalEdgeY;
-  }
-  if (isValidGridShift(drawer.gridShiftX) && drawer.gridShiftX !== 0) {
-    out.gridShiftX = drawer.gridShiftX;
-  }
-  if (isValidGridShift(drawer.gridShiftY) && drawer.gridShiftY !== 0) {
-    out.gridShiftY = drawer.gridShiftY;
-  }
-  if (drawer.outline !== undefined) {
-    out.outline = {
-      vertices: drawer.outline.vertices.map((v) =>
-        v.bulge === undefined || v.bulge === 0
-          ? { x: v.x, y: v.y }
-          : { x: v.x, y: v.y, bulge: v.bulge }
-      ),
-      ...(drawer.outline.authoring !== undefined &&
-      OUTLINE_AUTHORING_KINDS.includes(drawer.outline.authoring.kind)
-        ? {
-            authoring: {
-              kind: drawer.outline.authoring.kind,
-              // The corners editor round-trips its per-corner params through
-              // this annotation; keep them when every entry is structurally
-              // valid, drop the whole map otherwise.
-              ...(sanitizeCornerCuts(drawer.outline.authoring.corners) ?? {}),
-            },
-          }
-        : {}),
-    };
-  }
-  return out;
-}
-
-const CORNER_KEYS = ['tl', 'tr', 'bl', 'br'];
-const CUT_MAX_MM = 2600;
-
-function isValidCornerCut(value: unknown): value is CornerCutShape {
-  if (!isObject(value) || typeof value.kind !== 'string') return false;
-  switch (value.kind) {
-    case 'none':
-      return true;
-    case 'chamfer':
-      return isNumber(value.size) && inRange(value.size, 0, CUT_MAX_MM);
-    case 'radius':
-      return isNumber(value.r) && inRange(value.r, 0, CUT_MAX_MM);
-    case 'notch':
-      return (
-        isNumber(value.w) &&
-        inRange(value.w, 0, CUT_MAX_MM) &&
-        isNumber(value.d) &&
-        inRange(value.d, 0, CUT_MAX_MM)
-      );
-    default:
-      return false;
-  }
-}
-
-function sanitizeCornerCuts(corners: unknown): { corners: Record<string, CornerCutShape> } | null {
-  if (!isObject(corners)) return null;
-  const out: Record<string, CornerCutShape> = {};
-  for (const key of CORNER_KEYS) {
-    const cut = corners[key];
-    if (!isValidCornerCut(cut)) return null;
-    switch (cut.kind) {
-      case 'none':
-        out[key] = { kind: 'none' };
-        break;
-      case 'chamfer':
-        out[key] = { kind: 'chamfer', size: cut.size };
-        break;
-      case 'radius':
-        out[key] = { kind: 'radius', r: cut.r };
-        break;
-      default:
-        out[key] = { kind: 'notch', w: cut.w, d: cut.d };
-    }
-  }
-  return { corners: out };
 }
 
 function isValidLayer(value: unknown): value is LayerShape {
@@ -837,4 +401,3 @@ function sanitizeColor(color: string): string {
 }
 
 // Export constraints for use in other modules
-export { SHARE_CONSTRAINTS };
