+++ src/features/bin-designer/utils/printEstimates.ts
@@ -10,7 +10,7 @@
  * This avoids expensive mesh-based volume integration.
  */
 
-import type { BinParams, LabelTabSupport, WallPatternType } from '@/features/bin-designer/types';
+import type { BinParams } from '@/features/bin-designer/types';
 import {
   binFloorMm,
   isSocketlessBase,
@@ -18,16 +18,8 @@ import {
 } from '@/features/bin-designer/types/base';
 import { baseWallHeight } from './binDimensions';
 import { maxCompartmentFloorRaiseMm } from './compartmentFloorRaise';
-import {
-  DEFAULT_PATTERN_SCALE,
-  DEFAULT_PATTERN_WEB_THICKNESS,
-} from '@/features/bin-designer/types';
 import { GRIDFINITY, STYLE_WALL_THICKNESS } from '@/features/bin-designer/constants/gridfinity';
-import {
-  compartmentHasTiltedBackWall,
-  compartmentHasTiltedFrontWall,
-  getCompartmentBounds,
-} from '@/features/bin-designer/utils/compartments';
+import { getCompartmentBounds } from '@/features/bin-designer/utils/compartments';
 import { isFeatureActive } from '@/shared/constraints';
 import {
   PLA_DENSITY,
@@ -56,15 +48,16 @@ import {
   computeInteriorHeight,
   scoopFrameHeights,
 } from '@/shared/utils/scoopCalculations';
-import { resolveCompartmentDividerHeight } from '@/shared/utils/slotMath';
 import { countFilled, isPartialMask } from '@/shared/utils/cellMask';
-import { resolveWallPatternSides } from '@/shared/utils/wallPatternSides';
 import { cutoutDisplacementMm3 } from '@/shared/utils/fitTestPlan';
-import { FLOOR_PATTERN_BORDER, floorWindowSpan } from '@/shared/generation/floorPatternMetrics';
+import { computeLabelTabVolume, rampAreaWithin, lipSupportArea } from './printLabelTabVolume';
 import {
-  stampPatternOpenArea,
-  wallPatternElementMetrics,
-} from '@/shared/generation/wallPatternMetrics';
+  computeWallPatternReduction,
+  computeFloorPatternReduction,
+  effectiveDividerHeight,
+  totalDividerLength,
+} from './printPatternSavings';
+export { computeLabelTabVolume, rampAreaWithin, lipSupportArea } from './printLabelTabVolume';
 export interface PrintEstimate {
   /** Estimated material volume in mm³ */
   readonly volumeMm3: number;
@@ -347,25 +340,6 @@ function solidFillVolume(
 }
 
 /**
- * Height an interior divider actually reaches, in mm.
- *
- * Mirrors the generator: dividers rise from the cavity floor to the interior
- * ceiling (the stacking lip's bottom taper eats into it), capped by the
- * user's `dividerHeight`. Shared by the volume and the pattern-reduction terms
- * so shortening the dividers moves both together.
- */
-function effectiveDividerHeight(params: BinParams): number {
-  const totalH = params.height * params.heightUnitMm;
-  const wallHeight = baseWallHeight(params.base, totalH);
-  const interiorHeight = computeInteriorHeight(
-    wallHeight,
-    params.base.stackingLip,
-    GRIDFINITY.LIP_SMALL_TAPER
-  );
-  return resolveCompartmentDividerHeight(params.compartments.dividerHeight, interiorHeight);
-}
-
-/**
  * Volume of the slabs under raised compartments. Cell footprints rather than
  * cavity footprints: the divider insets are a couple of percent of a slab.
  */
@@ -405,275 +379,6 @@ function computeDividerVolume(
 
   // Volume = total wall length × thickness × height
   return totalDividerLength(params, innerW, innerD) * thickness * effectiveDividerHeight(params);
-}
-
-/** Summed length of every interior divider wall segment (mm). */
-function totalDividerLength(params: BinParams, innerW: number, innerD: number): number {
-  const { cols, rows, cells } = params.compartments;
-  if (cols <= 1 && rows <= 1) return 0;
-  const cellW = innerW / cols;
-  const cellD = innerD / rows;
-  let totalLength = 0;
-
-  for (let colBoundary = 1; colBoundary < cols; colBoundary++) {
-    for (let row = 0; row < rows; row++) {
-      const leftId = cells[row * cols + (colBoundary - 1)];
-      const rightId = cells[row * cols + colBoundary];
-      if (leftId !== rightId) {
-        totalLength += cellD;
-      }
-    }
-  }
-
-  for (let rowBoundary = 1; rowBoundary < rows; rowBoundary++) {
-    for (let col = 0; col < cols; col++) {
-      const topId = cells[(rowBoundary - 1) * cols + col];
-      const bottomId = cells[rowBoundary * cols + col];
-      if (topId !== bottomId) {
-        totalLength += cellW;
-      }
-    }
-  }
-
-  return totalLength;
-}
-/** Builder constant: max unsupported shelf span between bracket gussets. */
-const BRACKET_GUSSET_SPACING_MM = 10;
-
-interface LabelTabGeom {
-  readonly tabDepth: number;
-  readonly gussetLeg: number;
-  readonly wallThickness: number;
-  readonly dividerThickness: number;
-  readonly widthPercent: number;
-  readonly inset: number;
-  readonly support: LabelTabSupport;
-}
-
-/**
- * Volume of label tabs, mirroring `labelTabBuilder.buildTabsAtRow`'s
- * grouping and skip conditions (tilted anchor wall, depth+inset overrun,
- * `edges='both'` collisions) so the estimate tracks what the builder
- * actually generates.
- */
-function computeLabelTabVolume(
-  params: BinParams,
-  outerW: number,
-  outerD: number,
-  wallThickness: number
-): number {
-  const { cols, rows, thickness } = params.compartments;
-  const { depth: tabDepth, width: widthPercent, support } = params.label;
-  const inset = params.label.inset ?? 0;
-  const edges = params.label.edges ?? 'back';
-
-  const innerW = outerW - 2 * wallThickness;
-  const innerD = outerD - 2 * wallThickness;
-  const cellW = innerW / cols;
-  const cellD = innerD / rows;
-
-  // Mirrors the bridge guard in `buildLabelTabsInScope`.
-  if (tabDepth >= innerD) return 0;
-
-  // `gussetLeg` clamps at 0 so a tabDepth ≤ wallThickness yields shelf-only
-  // volume (no support), matching the builder's guard on degenerate geometry.
-  const geom: LabelTabGeom = {
-    tabDepth,
-    gussetLeg: Math.max(0, tabDepth - wallThickness),
-    wallThickness,
-    dividerThickness: thickness,
-    widthPercent,
-    inset,
-    support,
-  };
-
-  const includeBack = edges === 'back' || edges === 'both';
-  const includeFront = edges === 'front' || edges === 'both';
-  const collidingFrontIds =
-    edges === 'both' ? findCollidingFrontIds(params, cellD, tabDepth, inset) : null;
-
-  let volume = 0;
-  for (let row = 0; row < rows; row++) {
-    if (includeBack) {
-      volume += sumTabVolumesAtRow(params, row, 'back', cellW, cellD, geom, null);
-    }
-    if (includeFront) {
-      volume += sumTabVolumesAtRow(params, row, 'front', cellW, cellD, geom, collidingFrontIds);
-    }
-  }
-  return volume;
-}
-
-/**
- * Per-tab volume as a function of shelf width. Support depends on style:
- *   - `solid` / `fillet`: triangular prism extruded the full shelf width.
- *     Fillet has a concave profile but the over-estimate is small.
- *   - `bracket`: ~`ceil(tabWidth / BRACKET_GUSSET_SPACING_MM)` gussets,
- *     each extruded by the divider thickness (within ±1 of the builder).
- */
-function tabContribution(geom: LabelTabGeom, tabWidth: number): number {
-  if (tabWidth <= 0) return 0;
-  const shelf = geom.tabDepth * tabWidth * geom.wallThickness;
-  if (geom.gussetLeg <= 0) return shelf;
-  const triangleArea = 0.5 * geom.tabDepth * geom.gussetLeg;
-  const support =
-    geom.support === 'bracket'
-      ? Math.max(1, Math.ceil(tabWidth / BRACKET_GUSSET_SPACING_MM)) *
-        triangleArea *
-        geom.dividerThickness
-      : triangleArea * tabWidth;
-  return shelf + support;
-}
-
-/**
- * Sum tab volumes for one row + anchor edge. Mirrors the grouping and
- * skip logic of `labelTabBuilder.buildTabsAtRow`, including the
- * `thickness/2` boundary deduction at real divider walls.
- */
-function sumTabVolumesAtRow(
-  params: BinParams,
-  row: number,
-  anchor: 'back' | 'front',
-  cellW: number,
-  cellD: number,
-  geom: LabelTabGeom,
-  collidingFrontIds: Set<number> | null
-): number {
-  const { cols, rows, cells, thickness } = params.compartments;
-  const isOuterEdgeRow = anchor === 'back' ? row === rows - 1 : row === 0;
-  const neighborRowOffset = anchor === 'back' ? 1 : -1;
-  const hasTiltedAnchorWall =
-    anchor === 'back' ? compartmentHasTiltedBackWall : compartmentHasTiltedFrontWall;
-
-  const cellAt = (c: number): number => cells[row * cols + c];
-  const anchorsTab = (c: number): boolean =>
-    isOuterEdgeRow || cellAt(c) !== cells[(row + neighborRowOffset) * cols + c];
-
-  let volume = 0;
-  let col = 0;
-
-  while (col < cols) {
-    const cellId = cellAt(col);
-    if (!anchorsTab(col)) {
-      col++;
-      continue;
-    }
-    if (hasTiltedAnchorWall(params.compartments, cellId)) {
-      col++;
-      continue;
-    }
-    if (anchor === 'front' && collidingFrontIds?.has(cellId)) {
-      col++;
-      continue;
-    }
-
-    const bounds = getCompartmentBounds(params.compartments, cellId);
-    if (bounds) {
-      const compartmentDepth = (bounds.maxRow - bounds.minRow + 1) * cellD;
-      if (geom.tabDepth + geom.inset > compartmentDepth) {
-        col++;
-        continue;
-      }
-    }
-
-    let groupEnd = col + 1;
-    while (groupEnd < cols && cellAt(groupEnd) === cellId && anchorsTab(groupEnd)) {
-      groupEnd++;
-    }
-
-    const groupCols = groupEnd - col;
-    const groupMinCol = col;
-    const groupMaxCol = groupEnd - 1;
-
-    // Deduct half the divider thickness on either side that borders an
-    // actual divider wall (a different compartment). Outer bin walls don't
-    // deduct — the cellW already starts inside the bin wall.
-    const leftDeduction = groupMinCol > 0 && cellAt(groupMinCol - 1) !== cellId ? thickness / 2 : 0;
-    const rightDeduction =
-      groupMaxCol < cols - 1 && cellAt(groupMaxCol + 1) !== cellId ? thickness / 2 : 0;
-    const availableWidth = groupCols * cellW - leftDeduction - rightDeduction;
-    const tabWidth = (availableWidth * geom.widthPercent) / 100;
-
-    volume += tabContribution(geom, tabWidth);
-    col = groupEnd;
-  }
-
-  return volume;
-}
-
-/**
- * For `edges='both'`, identify compartments whose back+front tab pair would
- * collide so the front tab is dropped from the estimate. Mirror of
- * `labelTabBuilder.findCollidingFrontCompartments`.
- */
-function findCollidingFrontIds(
-  params: BinParams,
-  cellD: number,
-  tabDepth: number,
-  inset: number
-): Set<number> {
-  const { cols, rows, cells } = params.compartments;
-  const colliding = new Set<number>();
-  const visited = new Set<number>();
-
-  for (let row = 0; row < rows; row++) {
-    for (let col = 0; col < cols; col++) {
-      const cellId = cells[row * cols + col];
-      if (visited.has(cellId)) continue;
-      visited.add(cellId);
-
-      const bounds = getCompartmentBounds(params.compartments, cellId);
-      if (!bounds) continue;
-
-      const hasFrontAnchor =
-        bounds.minRow === 0 || cells[(bounds.minRow - 1) * cols + bounds.minCol] !== cellId;
-      const hasBackAnchor =
-        bounds.maxRow === rows - 1 || cells[(bounds.maxRow + 1) * cols + bounds.minCol] !== cellId;
-      if (!hasBackAnchor || !hasFrontAnchor) continue;
-
-      const compartmentDepth = (bounds.maxRow - bounds.minRow + 1) * cellD;
-      if (2 * tabDepth + 2 * inset > compartmentDepth) {
-        colliding.add(cellId);
-      }
-    }
-  }
-
-  return colliding;
-}
-
-/**
- * Material under one ramp's profile between the wall and `a` mm out along the
- * run: the part a divider's near half swallows when the ramp starts on its
- * centreline. Curved: the quarter-ellipse arc y = run(1 − cos θ),
- * z = height(1 − sin θ), so ∫z dy = run·height·[(1 − cos θa) − (θa/2 − sin 2θa/4)].
- */
-function rampAreaWithin(
-  profile: { run: number; height: number; style: string },
-  a: number
-): number {
-  const reach = Math.min(Math.max(0, a), profile.run);
-  if (profile.style !== 'curved') return profile.height * reach * (1 - reach / (2 * profile.run));
-  const theta = Math.acos(1 - reach / profile.run);
-  return (
-    profile.run * profile.height * (1 - Math.cos(theta) - (theta / 2 - Math.sin(2 * theta) / 4))
-  );
-}
-
-/**
- * Cross-section (mm²) of the stacking lip's angled support below the wall top:
- * the integrated lip's inner loft steps in from the wall face 2.65mm under the
- * top to the full overhang inset 1.2mm under it, then runs vertical. A lipped
- * outer ramp's strip climbs that same band flush with the lip base, so the
- * support's area is material the shell already prices. Mirrors
- * `integratedLipBuilder` (LIP_EXTENSION, FOOT: cross-feature import not allowed).
- */
-function lipSupportArea(wallThickness: number): number {
-  const LIP_EXTENSION = 1.2;
-  const FOOT = 0.05;
-  const top = Math.max(0, GRIDFINITY.LIP_SMALL_TAPER + GRIDFINITY.LIP_BIG_TAPER - wallThickness);
-  const foot = Math.max(0, LIP_EXTENSION - wallThickness);
-  const rampHeight = GRIDFINITY.LIP_SMALL_TAPER + GRIDFINITY.LIP_BIG_TAPER - LIP_EXTENSION - FOOT;
-  return ((foot + top) / 2) * rampHeight + top * LIP_EXTENSION;
 }
 
 /**
@@ -754,255 +459,6 @@ function computeScoopVolume(params: BinParams, outerW: number, outerD: number): 
     }
   }
   return volume;
-}
-/**
- * Approximate open-area fraction of each pattern at neutral scale — the share
- * of the wall face removed by the perforation. Honeycomb keeps its historical
- * 0.907 so existing estimates are unchanged; the sparser patterns remove less.
- *
- * Mirrors the geometry in patterns/* (cross-feature import not allowed). These
- * are estimates; the true fraction shifts slightly with web and keep-outs.
- */
-const PATTERN_VOID_FRACTION: Record<WallPatternType, number> = {
-  honeycomb: 0.907,
-  round: 0.62,
-  diamond: 0.72,
-  triangle: 0.42,
-  slots: 0.48,
-  mitsukude: 0.58,
-  goma: 0.45,
-  asanoha: 0.3,
-  sakura: 0.4,
-  rindo: 0.45,
-  mikado: 0.42,
-  'tsumiishi-kikko': 0.45,
-};
-
-/**
- * Volume removed by wall-pattern cutouts, for any pattern type.
- *
- * Approximation: pattern-specific open-area fraction of the wall face area,
- * modulated by the scale slider (bolder = more open, web fixed), cut through
- * the full wall thickness.
- */
-function computeWallPatternReduction(
-  params: BinParams,
-  outerW: number,
-  outerD: number,
-  totalH: number,
-  wallThickness: number,
-  bottomH: number
-): number {
-  // Must match wallPatterns.ts constants (cross-feature import not allowed)
-  const HEX_RADIUS = 1.8;
-  const WEB_THICKNESS = 0.8;
-  const TOP_KEEP_OUT = 1.5;
-  const BOTTOM_SOLID_SKIRT = 1.5;
-
-  const wallHeight = totalH - bottomH;
-  // wallThickness clears the floor slab; the skirt is the solid band above it
-  // that anchors the lowest element row. Mirrors wallPatterns.ts.
-  const bottomKeepOut = wallThickness + BOTTOM_SOLID_SKIRT;
-  const patternHeight = wallHeight - TOP_KEEP_OUT - bottomKeepOut;
-  const minPatternH = Math.sqrt(3) * HEX_RADIUS + WEB_THICKNESS;
-  if (patternHeight < minPatternH) return 0;
-
-  const innerW = outerW - 2 * wallThickness;
-  const innerD = outerD - 2 * wallThickness;
-
-  // Patterned wall length: slot-free walls (front/back seat y-axis dividers,
-  // left/right seat x-axis) intersected with the per-side selection.
-  const sides = resolveWallPatternSides(params.wallPattern);
-  const yFree = params.style !== 'slotted' || !params.slotConfig.y.enabled;
-  const xFree = params.style !== 'slotted' || !params.slotConfig.x.enabled;
-  let patternedWallLength = 0;
-  if (yFree) {
-    if (sides.front) patternedWallLength += innerW;
-    if (sides.back) patternedWallLength += innerW;
-  }
-  if (xFree) {
-    if (sides.left) patternedWallLength += innerD;
-    if (sides.right) patternedWallLength += innerD;
-  }
-
-  const wallFaceArea = patternedWallLength * patternHeight;
-
-  // Pattern open-area fraction, modulated by scale (0.5 = neutral, factor 1.0)
-  // and by the strut width, which the table's fractions assume at 0.8mm.
-  const scale = params.wallPattern.scale ?? DEFAULT_PATTERN_SCALE;
-  const base = PATTERN_VOID_FRACTION[params.wallPattern.pattern];
-  const coverageFraction = Math.min(
-    0.95,
-    Math.max(0, base * (0.85 + 0.3 * scale) * strutWidthFactor(params, patternHeight))
-  );
-  const coverage = wallFaceArea * coverageFraction;
-
-  // Material removed per unit area = wall thickness (prisms cut through wall)
-  const cutDepth = wallThickness;
-
-  return coverage * cutDepth + dividerPatternReduction(params, innerW, innerD, coverageFraction);
-}
-
-/**
- * How much of the default-web open area survives at this design's strut width.
- *
- * Read off the exact stamp model over a metre of band, where the edge margin
- * is noise, so the factor follows the real element spacing. 1 for a kumiko
- * lattice, which has no stamped web, and whenever the design keeps the
- * default; 0 when the wider strut lifts the pattern's minimum band height
- * past this wall, which is the case the worker leaves solid.
- */
-function strutWidthFactor(params: BinParams, patternHeight: number): number {
-  const { pattern, webThickness } = params.wallPattern;
-  if (webThickness === undefined || webThickness === DEFAULT_PATTERN_WEB_THICKNESS) return 1;
-  const scale = params.wallPattern.scale ?? DEFAULT_PATTERN_SCALE;
-  const { minPatternHeight } = wallPatternElementMetrics(
-    pattern,
-    params.height,
-    scale,
-    webThickness
-  );
-  if (patternHeight < minPatternHeight) return 0;
-  const reference = stampPatternOpenArea(pattern, params.height, scale, 1000, patternHeight);
-  if (reference <= 0) return 1;
-  const actual = stampPatternOpenArea(
-    pattern,
-    params.height,
-    scale,
-    1000,
-    patternHeight,
-    webThickness
-  );
-  return actual / reference;
-}
-
-/**
- * Extra material removed when the pattern is carried through the compartment
- * dividers.
- *
- * Same open-area model as the outer walls, applied to the divider face area
- * and cut through the divider's own thickness. The band is re-fitted to the
- * divider height, matching `planDividerPatterns` — a shortened divider offers
- * proportionally less to perforate.
- */
-function dividerPatternReduction(
-  params: BinParams,
-  innerW: number,
-  innerD: number,
-  coverageFraction: number
-): number {
-  // Mirrors the worker's `dividerPatternsApply` gate — subtracting for a bin
-  // the generator never patterns would make the estimate disagree with the
-  // geometry it is describing.
-  if (params.wallPattern.dividers !== true) return 0;
-  if (params.style !== 'standard') return 0;
-  if (params.base.solid) return 0;
-  if (isPartialMask(params.cellMask)) return 0;
-  if (params.compartments.thickness <= 0) return 0;
-
-  const dividerHeight = effectiveDividerHeight(params);
-
-  // Mirrors wallPatterns.ts (cross-feature import not allowed).
-  const TOP_KEEP_OUT = 1.5;
-  const BOTTOM_SOLID_SKIRT = 1.5;
-  const bandHeight = dividerHeight - TOP_KEEP_OUT - (params.wallThickness + BOTTOM_SOLID_SKIRT);
-  if (bandHeight <= 0) return 0;
-
-  const length = totalDividerLength(params, innerW, innerD);
-  if (length <= 0) return 0;
-
-  return length * bandHeight * coverageFraction * params.compartments.thickness;
-}
-/**
- * Volume removed by the floor pattern.
- *
- * Two cut depths, priced two ways. A SOCKETED integral base is cut straight
- * through feet and slab, so the removal is the open share of the base
- * component (`standardBinSolidComponents.base`, solid-calibrated per #3528 —
- * a foot is not a prism, so a share of the measured term beats an area×depth
- * model). Every SLAB-ONLY base — flat, the underside relief (whose cavity is
- * already open below the slab) and detachable feet (whose body floor sits at
- * Z=0) — only loses `wallThickness` of material under each hole, so the
- * removal is the open area times the slab, priced directly. Booking the
- * full-socket share for those cut a 3x3 estimate roughly in half for a
- * pattern that removes ~4mm³ per cm² of window.
- *
- * Open area is measured from the real element placement rather than a
- * fraction-of-area model: a per-foot window is only ~33mm across, and over a box
- * that small the margin the calculator leaves at every edge is most of the box.
- *
- * Mirrors the worker's `floorPatternApplies` gate so the estimate can't claim a
- * saving the generator never makes.
- */
-function computeFloorPatternReduction(
-  params: BinParams,
-  wallThickness: number,
-  baseVolume: number,
-  feetDetach: boolean
-): number {
-  const floorPattern = params.floorPattern;
-  if (floorPattern?.enabled !== true) return 0;
-  // `style === 'solid'` alongside `base.solid`: the two are kept in lockstep by
-  // a runtime IMPLICATION_RULE, not by `migrateParams`, so a crafted design can
-  // carry one without the other.
-  if (params.base.solid || params.style === 'solid') return 0;
-  // Mirrors `floorPatternApplies`: an interior lite floor has no slab left to
-  // perforate, but the underside relief keeps one and still cuts the holes.
-  if (params.base.lightweight && !isUndersideRelief(params.base)) return 0;
-  if (params.base.spacer) return 0;
-  // The generator forces a base-only bin down the solid path, so no floor
-  // pattern is ever cut. `base.solid` above does not catch it: IMPLICATION_RULES
-  // hold that flag false while the tray keeps style 'standard'.
-  if (params.base.tile === true) return 0;
-  if (params.width <= 0 || params.depth <= 0) return 0;
-
-  const isFlat = isSocketlessBase(params.base.style);
-  const gridUnitMmY = params.gridUnitMmY ?? params.gridUnitMm;
-  const scale = floorPattern.scale ?? DEFAULT_PATTERN_SCALE;
-  const openArea = (w: number, h: number): number =>
-    stampPatternOpenArea(floorPattern.pattern, params.height, scale, w, h);
-
-  const outerW = params.width * params.gridUnitMm - GRIDFINITY.TOLERANCE;
-  const outerD = params.depth * gridUnitMmY - GRIDFINITY.TOLERANCE;
-  // A flat base has no feet to thread the holes through, so the whole cavity
-  // floor is one window; a socket bin gets one window per foot — and
-  // `halfSockets` quarters every foot, so each cell carries four much smaller
-  // windows rather than one big one. Assuming full-cell windows there would
-  // over-report the open area several times over, since a window that small
-  // fits disproportionately fewer elements.
-  const cellUnits = params.base.halfSockets ? 0.5 : 1;
-  const windowsPerCell = params.base.halfSockets ? 4 : 1;
-  // A custom footprint only grows feet on its filled cells, so only those carry
-  // windows. Scaling by the filled fraction tracks that without re-deriving the
-  // socket builder's per-cell decomposition here — and it cancels the fact that
-  // `baseVolume` counts the full bounding rectangle, leaving the reduction
-  // proportional to the cells that actually exist.
-  const filledFraction = isPartialMask(params.cellMask)
-    ? countFilled(params.cellMask) / (params.cellMask.cols * params.cellMask.rows)
-    : 1;
-  const totalOpenArea = isFlat
-    ? openArea(
-        outerW - 2 * wallThickness - 2 * FLOOR_PATTERN_BORDER,
-        outerD - 2 * wallThickness - 2 * FLOOR_PATTERN_BORDER
-      )
-    : params.width *
-      params.depth *
-      filledFraction *
-      windowsPerCell *
-      openArea(
-        floorWindowSpan(cellUnits, params.gridUnitMm, params.wallThickness),
-        floorWindowSpan(cellUnits, gridUnitMmY, params.wallThickness)
-      );
-  if (totalOpenArea <= 0) return 0;
-
-  const planArea = params.width * params.depth * params.gridUnitMm * gridUnitMmY;
-  // See the docstring: slab-only cuts are priced as the direct prism, the
-  // socketed full-depth cut as a share of the measured base term.
-  const slabOnly = isFlat || isUndersideRelief(params.base) || feetDetach;
-  if (slabOnly) {
-    return Math.min(totalOpenArea, planArea) * wallThickness;
-  }
-  return baseVolume * Math.min(1, totalOpenArea / planArea);
 }
 
 export interface WallPatternSavings {
